Add browserify
