+++ src/index.js
@@ -1,41 +1,4 @@
-'use strict';
-
-/* eslint-env browser */
-
-const input = document.querySelector('.Plugin-filter-input');
-const plugins = [];
-
-/**
-* Filter plugins.
-*/
-
-const filter = () => {
-  let value = input.value.toLowerCase();
-
-  plugins.forEach((plugin) => {
-    const { el, title, description } = plugin;
-
-    el.style.display = '';
-
-    if (!value) {
-      return;
-    }
-    if (!~title.indexOf(value) && !~description.indexOf(value)) {
-      el.style.display = 'none';
-    }
-  });
-};
-
-/**
- * Build index.
- */
-Array.from(document.querySelectorAll('.Plugin-list .Plugin')).forEach((el) => {
-  plugins.push({
-    el,
-    title: el.querySelector('.Plugin-title').textContent.toLowerCase(),
-    description: el.querySelector('.Plugin-description').textContent.toLowerCase()
-  });
-});
+const filter = require('../lib/js/filter');
 
 /**
  * Filter to start, in case there is some text in the search input, which may
@@ -43,5 +6,7 @@ Array.from(document.querySelectorAll('.Plugin-list .Plugin')).forEach((el) => {
  * future filtering.
  */
 
-filter();
-input.addEventListener('keyup', filter);
+const input = document.querySelector('.Plugin-filter-input');
+
+filter(input);
+input.addEventListener('keyup', () => filter(input));
